pyproject.toml
[project]
name = "lasse-blog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0",
    "gunicorn>=23.0.0",
    "psycopg>=3.3.1",
    "pymemcache>=4.0.0",
    "wagtail>=7.2.1",
    "wagtailcodeblock>=1.30.0.0",
]

[dependency-groups]
dev = [
    "debugpy>=1.8.19",
    "django-browser-reload>=1.21.0",
    "django-debug-toolbar>=6.1.0",
    "django-pattern-library>=1.5.0",
    "django-upgrade>=1.29.1",
    "djhtml>=3.0.10",
    "honcho>=2.0.0",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-django>=4.11.1",
    "ruff>=0.14.10",
    "time-machine>=3.2.0",
    "wagtail-factories>=4.3.0",
]
playwright = [
    "pytest-playwright>=0.7.2",
]


[tool.ruff]
extend-exclude = ["**/migrations/*.py"]


[tool.pytest]
addopts = ["--reuse-db"]
DJANGO_SETTINGS_MODULE = "blog.settings.test"
pythonpath = ["src"]
python_files = ["**/tests/test_*.py", "**/functional_tests/test_*.py"]
